refactor: move BitbucketPullRequestState and BitbucketIssueState into IBitbucketService.ts
