+++ src/services/bitbucket/IBitbucketService.ts
@@ -1,0 +1,14 @@
+export enum BitbucketPullRequestState {
+  open = 'OPEN',
+  closed = 'MERGED',
+  declined = 'DECLINED',
+  superseded = 'SUPERSEDED',
+}
+
+export enum BitbucketIssueState {
+  new = 'new',
+  closed = 'closed',
+  duplicate = 'duplicate',
+  resolved = 'resolved',
+  invalid = 'invalid',
+}
